Deny trivial casts and fix them all

+++ src/zopfli/squeeze.rs
@@ -115,7 +115,7 @@ impl SymbolStats {
         fn randomize_freqs(freqs: &mut [size_t], state: &mut RanState) {
             let n = freqs.len();
             let mut i: usize = 0;
-            let end = n as usize;
+            let end = n;
 
             while i < end {
                 if (state.random_marsaglia() >> 4) % 3 == 0 {
@@ -357,7 +357,7 @@ pub fn get_best_lengths(s: &mut ZopfliBlockState, in_data: *mut c_uchar, instart
     let mut length_array = vec![0; blocksize + 1];
     let mut leng;
     let mut longest_match;
-    let sublen = unsafe { malloc(mem::size_of::<c_ushort>() as size_t * 259) as *mut c_ushort };
+    let sublen = unsafe { malloc(mem::size_of::<c_ushort>() * 259) as *mut c_ushort };
     let windowstart = if instart > ZOPFLI_WINDOW_SIZE {
         instart - ZOPFLI_WINDOW_SIZE
     } else {
